src(scales): move scale_utils/ to utils/scale_utils
