--- cellengine/payloads/scale_utils/scale_dict.py
@@ -1,5 +1,0 @@
-from cellengine.utils.limited_dict import LimitedDict
-
-
-class ScaleDict(LimitedDict):
-    _keys = {"type", "minimum", "maximum", "cofactor"}
