+++ tests/unit/resources/scales/test_linear_scale.py
@@ -1,7 +1,7 @@
 import pytest
 from pandas import Series
 
-from cellengine.payloads.scale_utils.apply_scale import apply_scale
+from cellengine.utils.scale_utils import apply_scale
 
 
 @pytest.fixture(scope="module")
